Atskyriau PostTagsController. Pridejau i Room modeli virtual fielda tagams kuri buvau pamirses. Ikeliau pradinius tag ir posttags migrationus jei ju prireiks.
